refactor(application): extract dependency_builder.rs and license_compliance_builder.rs [#314-3/5]
Extract build_dependencies and build_license_compliance from mod.rs
into dedicated sub-builder files as part of #314.

- Create dependency_builder.rs with pub(super) build_dependencies
- Create license_compliance_builder.rs with pub(super) build_license_compliance
- Update mod.rs to delegate to the new sub-builders
- Add HashMap import in test module to support existing tests

Closes #339
